pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = "Тестовое задание в Московский Метрополитен"
authors = ["Elisha Nimas"]
readme = "README.md"
packages = [{ include = "ocr_service", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.110.0"
uvicorn = { version = "^0.27.1", extras = ["standard"] }
easyocr = "^1.7.1"
opencv-python-headless = "^4.9.0.80"
pillow = "^10.2.0"
numpy = "^1.26.4"
transformers = "^4.40.0"
torch = "^2.2.0"
sentencepiece = "^0.2.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
pytest = "^8.0.0"
types-pillow = "^10.2.0.20240213"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
